Created the Break Time Resource

+++ src/main/resources/db/migration/V5__Create_Break_Time_Table.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS BreakTime(
     employee_id binary(16) NOT NULL,
     break_start DATETIME NOT NULL,
     break_end DATETIME NOT NULL,
-    status BIGINT NOT NULL DEFAULT 0,
+    status TINYINT NOT NULL DEFAULT 0,
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     modified_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (employee_id) REFERENCES Employee(employee_id)
